+++ crates/validator-debt/CHANGELOG.md
@@ -6,7 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
+ - fetch deposit accounts ([#]())
+## [] - 
 - use vote key from past ([#250](https://github.com/doublezerofoundation/doublezero-offchain/pull/250))
 - bail early when Revenue Distribution program is paused ([#244](https://github.com/doublezerofoundation/doublezero-offchain/pull/244))
 - finalize zero debt ([#248](https://github.com/doublezerofoundation/doublezero-offchain/pull/248))